RW-607 fixed test.
