+++ api/db-cdr/changelog-data/db.changelog-16.xml
@@ -7,6 +7,7 @@
         http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd">
   <changeSet author="srushtigangireddy" id="changelog-16" runAlways="true" context="local or cloud">
 
+    <validCheckSum>ANY</validCheckSum>
 
     <delete tableName="db_domain"/>
