+++ api/src/test/java/org/pmiops/workbench/api/CohortReviewControllerTest.java
@@ -1,410 +1,310 @@
 package org.pmiops.workbench.api;
 
-import com.google.cloud.bigquery.FieldValue;
-import com.google.cloud.bigquery.QueryResult;
-import com.google.gson.Gson;
+import static com.google.common.truth.Truth.assertThat;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.pmiops.workbench.cdr.CdrVersionContext;
 import org.pmiops.workbench.cdr.cache.GenderRaceEthnicityConcept;
+import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
 import org.pmiops.workbench.cohortbuilder.ParticipantCounter;
-import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
-import org.pmiops.workbench.cohortreview.CohortReviewService;
+import org.pmiops.workbench.cohortreview.CohortReviewServiceImpl;
+import org.pmiops.workbench.cohortreview.ReviewTabQueryBuilder;
+import org.pmiops.workbench.db.dao.CdrVersionDao;
+import org.pmiops.workbench.db.dao.CohortDao;
+import org.pmiops.workbench.db.dao.CohortReviewDao;
+import org.pmiops.workbench.db.dao.ParticipantCohortStatusDao;
+import org.pmiops.workbench.db.dao.WorkspaceDao;
 import org.pmiops.workbench.db.dao.WorkspaceService;
+import org.pmiops.workbench.db.model.CdrVersion;
 import org.pmiops.workbench.db.model.Cohort;
-import org.pmiops.workbench.db.model.CohortAnnotationDefinition;
-import org.pmiops.workbench.db.model.CohortAnnotationEnumValue;
 import org.pmiops.workbench.db.model.CohortReview;
 import org.pmiops.workbench.db.model.ParticipantCohortStatus;
 import org.pmiops.workbench.db.model.ParticipantCohortStatusKey;
 import org.pmiops.workbench.db.model.Workspace;
-import org.pmiops.workbench.exceptions.BadRequestException;
-import org.pmiops.workbench.exceptions.NotFoundException;
-import org.pmiops.workbench.model.*;
+import org.pmiops.workbench.model.CohortStatus;
+import org.pmiops.workbench.model.DataAccessLevel;
+import org.pmiops.workbench.model.PageFilterRequest;
+import org.pmiops.workbench.model.ParticipantCohortStatusColumns;
+import org.pmiops.workbench.model.ParticipantCohortStatuses;
+import org.pmiops.workbench.model.SortOrder;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.liquibase.LiquibaseAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.junit4.SpringRunner;
 
-import javax.inject.Provider;
-import java.sql.Date;
 import java.sql.Timestamp;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeSet;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@RunWith(SpringRunner.class)
+@DataJpaTest
+@Import(LiquibaseAutoConfiguration.class)
+@AutoConfigureTestDatabase(replace= AutoConfigureTestDatabase.Replace.NONE)
 public class CohortReviewControllerTest {
 
-    private String namespace;
-    private String name;
-    private long cohortId;
-    private long cdrVersionId;
-    private long workspaceId;
-    private long cohortReviewId;
-    private long participantId;
+  private static final Long COHORT_ID = 1L;
+  private static final Long WORKSPACE_ID = 1L;
+  private static final String WORKSPACE_NAMESPACE = "namespace";
+  private static final String WORKSPACE_NAME = "name";
+  private static final Long CDR_VERSION_ID = 1L;
+  private CohortReview cohortReview;
+  private ParticipantCohortStatus participantCohortStatus1;
+  private ParticipantCohortStatus participantCohortStatus2;
 
-    @Mock
-    CohortReviewService cohortReviewService;
+  private enum TestDemo {
+    ASIAN("Asian", 8515),
+    WHITE("White", 8527),
+    MALE("MALE", 8507),
+    FEMALE("FEMALE", 8532),
+    NOT_HISPANIC("Not Hispanic or Latino", 38003564);
 
-    @Mock
-    BigQueryService bigQueryService;
+    private final String name;
+    private final long conceptId;
 
-    @Mock
-    ParticipantCounter participantCounter;
-
-    @Mock
-    DomainLookupService domainLookupService;
-
-    @Mock
-    WorkspaceService workspaceService;
-
-    @Mock
-    Provider<GenderRaceEthnicityConcept> genderRaceEthnicityConceptProvider;
-
-    @InjectMocks
-    CohortReviewController reviewController;
-
-    @Before
-    public void onSetUp() {
-        namespace = "aou-test";
-        name = "test";
-        cohortId = 1;
-        cdrVersionId = 1;
-        workspaceId = 1;
-        cohortReviewId = 1;
-        participantId = 1;
+    private TestDemo(String name, long conceptId) {
+      this.name = name;
+      this.conceptId = conceptId;
     }
 
-    @Test
-    public void createCohortReviewReviewAlreadyCreated() throws Exception {
-
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(1, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(createWorkspace(workspaceId, namespace, name));
-
-        try {
-            reviewController.createCohortReview(namespace, name, cohortId, cdrVersionId, new CreateReviewRequest().size(200));
-            fail("Should have thrown a BadRequestException!");
-        } catch (BadRequestException e) {
-            assertEquals("Invalid Request: Cohort Review already created for cohortId: "
-                    + cohortId + ", cdrVersionId: " + cdrVersionId, e.getMessage());
-        }
-
-        verify(cohortReviewService, times(1)).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService, times(1)).findCohort(cohortId);
-        verify(cohortReviewService, times(1)).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verifyNoMoreMockInteractions();
+    public String getName() {
+      return name;
     }
 
-    @Test
-    public void createCohortReviewMoreThanTenThousand() throws Exception {
-
-        try {
-            reviewController.createCohortReview(namespace, name, cohortId, cdrVersionId, new CreateReviewRequest().size(20000));
-            fail("Should have thrown a BadRequestException!");
-        } catch (BadRequestException e) {
-            assertEquals("Invalid Request: Cohort Review size must be between 0 and 10000", e.getMessage());
-        }
-
-        verifyNoMoreMockInteractions();
+    public long getConceptId() {
+      return conceptId;
     }
+  }
 
-    @Test
-    public void createCohortReviewNoCohortDefinitionFound() throws Exception {
+  @TestConfiguration
+  @Import({
+    CohortReviewController.class,
+    CohortReviewServiceImpl.class,
+    ParticipantCounter.class,
+    CohortQueryBuilder.class,
+    DomainLookupService.class,
+    ReviewTabQueryBuilder.class
+  })
+  @MockBean({
+    WorkspaceService.class,
+    BigQueryService.class
+  })
+  static class Configuration {
 
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(createWorkspace(workspaceId, namespace, name));
-
-        try {
-            reviewController.createCohortReview(namespace, name, cohortId, cdrVersionId, new CreateReviewRequest().size(200));
-            fail("Should have thrown a NotFoundException!");
-        } catch (NotFoundException e) {
-            assertEquals("Not Found: No Cohort definition matching cohortId: "
-                         + cohortId,
-                         e.getMessage());
-        }
-
-        verify(cohortReviewService, times(1)).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService, times(1)).findCohort(cohortId);
-        verify(cohortReviewService, times(1)).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verifyNoMoreMockInteractions();
+    @Bean
+    GenderRaceEthnicityConcept getGenderRaceEthnicityConcept() {
+      Map<String, Map<Long, String>> concepts = new HashMap<>();
+      concepts.put(ParticipantCohortStatusColumns.RACE.name(),
+        new HashMap<Long, String>() {{
+          put(TestDemo.ASIAN.getConceptId(), TestDemo.ASIAN.getName());
+          put(TestDemo.WHITE.getConceptId(), TestDemo.WHITE.getName()); }});
+      concepts.put(ParticipantCohortStatusColumns.GENDER.name(),
+        new HashMap<Long, String>() {{
+          put(TestDemo.MALE.getConceptId(), TestDemo.MALE.getName());
+          put(TestDemo.FEMALE.getConceptId(), TestDemo.FEMALE.getName()); }});
+      concepts.put(ParticipantCohortStatusColumns.ETHNICITY.name(),
+        new HashMap<Long, String>() {{
+          put(TestDemo.NOT_HISPANIC.getConceptId(), TestDemo.NOT_HISPANIC.getName()); }});
+      return new GenderRaceEthnicityConcept(concepts);
     }
+  }
 
-    @Test
-    public void createCohortReview() throws Exception {
+  private Workspace workspace;
 
-        String definition = "{\"includes\":[{\"items\":[{\"type\":\"DEMO\",\"searchParameters\":" +
-                "[{\"value\":\"Age\",\"subtype\":\"AGE\",\"conceptId\":null,\"attribute\":" +
-                "{\"operator\":\"between\",\"operands\":[18,66]}}],\"modifiers\":[]}]}],\"excludes\":[]}";
+  @Autowired
+  private CdrVersionDao cdrVersionDao;
 
-        SearchRequest searchRequest = new Gson().fromJson(definition, SearchRequest.class);
+  @Autowired
+  private WorkspaceDao workspaceDao;
 
-        QueryResult queryResult = mock(QueryResult.class);
-        Iterable testIterable = new Iterable() {
-            @Override
-            public Iterator iterator() {
-                List<FieldValue> list = new ArrayList<>();
-                list.add(null);
-                return list.iterator();
-            }
-        };
-        Map<String, Integer> rm = new HashMap<>();
-        rm.put("person_id", 0);
-        rm.put("birth_datetime", 1);
-        rm.put("gender_concept_id", 2);
-        rm.put("race_concept_id", 3);
-        rm.put("ethnicity_concept_id", 4);
+  @Autowired
+  private CohortDao cohortDao;
 
-        when(workspaceService.enforceWorkspaceAccessLevel(namespace, name, WorkspaceAccessLevel.READER)).thenReturn(WorkspaceAccessLevel.OWNER);
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, definition));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(createWorkspace(workspaceId, namespace, name));
-        when(participantCounter.buildParticipantIdQuery(new ParticipantCriteria(searchRequest), 200, 0L)).thenReturn(null);
-        when(bigQueryService.filterBigQueryConfig(null)).thenReturn(null);
-        when(bigQueryService.executeQuery(null)).thenReturn(queryResult);
-        when(bigQueryService.getResultMapper(queryResult)).thenReturn(rm);
-        when(queryResult.iterateAll()).thenReturn(testIterable);
-        when(bigQueryService.getLong(null, 0)).thenReturn(0L);
-        when(bigQueryService.getString(null, 1)).thenReturn("1");
-        when(bigQueryService.getLong(null, 2)).thenReturn(0L);
-        when(bigQueryService.getLong(null, 3)).thenReturn(0L);
-        when(bigQueryService.getLong(null, 4)).thenReturn(0L);
-        when(genderRaceEthnicityConceptProvider.get()).thenReturn(new GenderRaceEthnicityConcept(createGenderRaceEthnicityConcept()));
-        doNothing().when(cohortReviewService).saveFullCohortReview(createCohortReview(1, cohortId, cohortReviewId, cdrVersionId, ReviewStatus.CREATED),
-                Arrays.asList(createParticipantCohortStatus(cohortReviewId, 0, CohortStatus.NOT_REVIEWED)));
-        when(cohortReviewService.findAll(isA(Long.class), isA(List.class), isA(PageRequest.class)))
-                .thenReturn(Arrays.asList(createParticipantCohortStatus(cohortReviewId, 0, CohortStatus.INCLUDED)));
+  @Autowired
+  private CohortReviewDao cohortReviewDao;
 
-        reviewController.createCohortReview(namespace, name, cohortId, cdrVersionId, new CreateReviewRequest().size(200));
+  @Autowired
+  private ParticipantCohortStatusDao participantCohortStatusDao;
 
-        verify(cohortReviewService, times(1)).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService, times(1)).findCohort(cohortId);
-        verify(cohortReviewService, times(1)).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verify(participantCounter, times(1)).buildParticipantIdQuery(new ParticipantCriteria(searchRequest), 200, 0L);
-        verify(bigQueryService, times(1)).filterBigQueryConfig(null);
-        verify(bigQueryService, times(1)).executeQuery(null);
-        verify(bigQueryService, times(1)).getResultMapper(queryResult);
-        verify(bigQueryService, times(1)).getLong(null, 0);
-        verify(bigQueryService, times(1)).getString(null, 1);
-        verify(bigQueryService, times(1)).getLong(null, 2);
-        verify(bigQueryService, times(1)).getLong(null, 3);
-        verify(bigQueryService, times(1)).getLong(null, 4);
-        verify(queryResult, times(1)).iterateAll();
-        verify(cohortReviewService, times(1)).saveFullCohortReview(isA(CohortReview.class), isA(List.class));
-        verify(genderRaceEthnicityConceptProvider, times(1)).get();
-        verify(cohortReviewService).findAll(isA(Long.class), isA(List.class), isA(PageRequest.class));
-        verifyNoMoreMockInteractions();
+  @Autowired
+  private WorkspaceService workspaceService;
+
+  @Autowired
+  private CohortReviewController cohortReviewController;
+
+  @Before
+  public void setUp() {
+    CdrVersion cdrVersion = new CdrVersion();
+    cdrVersion.setBigqueryDataset("dataSetId");
+    cdrVersion.setBigqueryProject("projectId");
+    cdrVersionDao.save(cdrVersion);
+    CdrVersionContext.setCdrVersion(cdrVersion);
+
+    workspace = new Workspace();
+    workspace.setCdrVersion(cdrVersion);
+    workspace.setName("name");
+    workspace.setDataAccessLevel(DataAccessLevel.PROTECTED);
+    workspaceDao.save(workspace);
+
+    Cohort cohort = new Cohort();
+    cohort.setCohortId(COHORT_ID);
+    cohort.setWorkspaceId(WORKSPACE_ID);
+    cohortDao.save(cohort);
+
+    Timestamp today = new Timestamp(new Date().getTime());
+    cohortReview = cohortReviewDao.save(
+      new CohortReview()
+      .cohortId(COHORT_ID)
+      .cdrVersionId(CDR_VERSION_ID))
+      .creationTime(today);
+
+    ParticipantCohortStatusKey key1 = new ParticipantCohortStatusKey()
+      .cohortReviewId(cohortReview.getCohortReviewId())
+      .participantId(1L);
+    ParticipantCohortStatusKey key2 = new ParticipantCohortStatusKey()
+      .cohortReviewId(cohortReview.getCohortReviewId())
+      .participantId(2L);
+
+    participantCohortStatus1 = new ParticipantCohortStatus()
+      .status(CohortStatus.NOT_REVIEWED)
+      .participantKey(key1)
+      .genderConceptId(TestDemo.MALE.getConceptId())
+      .gender(TestDemo.MALE.getName())
+      .raceConceptId(TestDemo.ASIAN.getConceptId())
+      .race(TestDemo.ASIAN.getName())
+      .ethnicityConceptId(TestDemo.NOT_HISPANIC.getConceptId())
+      .ethnicity(TestDemo.NOT_HISPANIC.getName())
+      .birthDate(new java.sql.Date(today.getTime()));
+    participantCohortStatus2 = new ParticipantCohortStatus()
+      .status(CohortStatus.NOT_REVIEWED)
+      .participantKey(key2)
+      .genderConceptId(TestDemo.FEMALE.getConceptId())
+      .gender(TestDemo.FEMALE.getName())
+      .raceConceptId(TestDemo.WHITE.getConceptId())
+      .race(TestDemo.WHITE.getName())
+      .ethnicityConceptId(TestDemo.NOT_HISPANIC.getConceptId())
+      .ethnicity(TestDemo.NOT_HISPANIC.getName())
+      .birthDate(new java.sql.Date(today.getTime()));
+
+    participantCohortStatusDao.save(participantCohortStatus1);
+    participantCohortStatusDao.save(participantCohortStatus2);
+  }
+
+  @Test
+  public void getParticipantCohortStatuses() throws Exception {
+    int page = 0;
+    int pageSize = 25;
+    org.pmiops.workbench.model.CohortReview expectedReview1 =
+      createCohortReview(cohortReview,
+        Arrays.asList(participantCohortStatus1, participantCohortStatus2),
+        page,
+        pageSize,
+        SortOrder.DESC,
+        ParticipantCohortStatusColumns.STATUS);
+    org.pmiops.workbench.model.CohortReview expectedReview2 =
+      createCohortReview(cohortReview,
+        Arrays.asList(participantCohortStatus2, participantCohortStatus1),
+        page,
+        pageSize,
+        SortOrder.DESC,
+        ParticipantCohortStatusColumns.PARTICIPANTID);
+    org.pmiops.workbench.model.CohortReview expectedReview3 =
+      createCohortReview(cohortReview,
+        Arrays.asList(participantCohortStatus1, participantCohortStatus2),
+        page,
+        pageSize,
+        SortOrder.ASC,
+        ParticipantCohortStatusColumns.STATUS);
+    org.pmiops.workbench.model.CohortReview expectedReview4 =
+      createCohortReview(cohortReview,
+        Arrays.asList(participantCohortStatus1, participantCohortStatus2),
+        page,
+        pageSize,
+        SortOrder.ASC,
+        ParticipantCohortStatusColumns.PARTICIPANTID);
+
+    when(workspaceService.getRequired(WORKSPACE_NAMESPACE, WORKSPACE_NAME)).thenReturn(workspace);
+
+    assertParticipantCohortStatuses(expectedReview1, page, pageSize, SortOrder.DESC, ParticipantCohortStatusColumns.STATUS);
+    assertParticipantCohortStatuses(expectedReview2, page, pageSize, SortOrder.DESC, ParticipantCohortStatusColumns.PARTICIPANTID);
+    assertParticipantCohortStatuses(expectedReview3, null, null, null, ParticipantCohortStatusColumns.STATUS);
+    assertParticipantCohortStatuses(expectedReview4, null, null, SortOrder.ASC, null);
+    assertParticipantCohortStatuses(expectedReview4, null, pageSize, null, null);
+    assertParticipantCohortStatuses(expectedReview4, page, null, null, null);
+    assertParticipantCohortStatuses(expectedReview4, null, null, null, null);
+  }
+
+  /**
+   * Helper method to assert results for
+   * {@link CohortReviewController#getParticipantCohortStatuses(String, String, Long, Long, PageFilterRequest)}.
+   *
+   * @param expectedReview
+   * @param page
+   * @param pageSize
+   * @param sortOrder
+   * @param sortColumn
+   */
+  private void assertParticipantCohortStatuses(org.pmiops.workbench.model.CohortReview expectedReview,
+                                               Integer page,
+                                               Integer pageSize,
+                                               SortOrder sortOrder,
+                                               ParticipantCohortStatusColumns sortColumn) {
+    org.pmiops.workbench.model.CohortReview actualReview =
+      cohortReviewController.getParticipantCohortStatuses(WORKSPACE_NAMESPACE,
+      WORKSPACE_NAME,
+      COHORT_ID,
+      CDR_VERSION_ID,
+        new ParticipantCohortStatuses()
+          .sortColumn(sortColumn)
+          .page(page)
+          .pageSize(pageSize)
+          .sortOrder(sortOrder)).getBody();
+
+    assertThat(actualReview).isEqualTo(expectedReview);
+  }
+
+  private org.pmiops.workbench.model.CohortReview createCohortReview(CohortReview actualReview,
+                                                                     List<ParticipantCohortStatus> participantCohortStatusList,
+                                                                     Integer page,
+                                                                     Integer pageSize,
+                                                                     SortOrder sortOrder,
+                                                                     ParticipantCohortStatusColumns sortColumn) {
+    List<org.pmiops.workbench.model.ParticipantCohortStatus> newParticipantCohortStatusList = new ArrayList<>();
+    for (ParticipantCohortStatus participantCohortStatus : participantCohortStatusList) {
+      newParticipantCohortStatusList.add(new org.pmiops.workbench.model.ParticipantCohortStatus()
+        .birthDate(participantCohortStatus.getBirthDate().toString())
+        .ethnicityConceptId(participantCohortStatus.getEthnicityConceptId())
+        .ethnicity(participantCohortStatus.getEthnicity())
+        .genderConceptId(participantCohortStatus.getGenderConceptId())
+        .gender(participantCohortStatus.getGender())
+        .participantId(participantCohortStatus.getParticipantKey().getParticipantId())
+        .raceConceptId(participantCohortStatus.getRaceConceptId())
+        .race(participantCohortStatus.getRace())
+        .status(participantCohortStatus.getStatus()));
     }
-
-    @Test
-    public void createParticipantCohortAnnotation() throws Exception {
-        assertCreateParticipantCohortAnnotation(createParticipantCohortAnnotation(Boolean.TRUE, null, null, null, null), AnnotationType.BOOLEAN);
-        assertCreateParticipantCohortAnnotation(createParticipantCohortAnnotation(null, "test", null, null, null), AnnotationType.STRING);
-        assertCreateParticipantCohortAnnotation(createParticipantCohortAnnotation(null, null, 1, null, null), AnnotationType.INTEGER);
-        assertCreateParticipantCohortAnnotation(createParticipantCohortAnnotation(null, null, null, "1999-02-01", null), AnnotationType.DATE);
-        assertCreateParticipantCohortAnnotation(createParticipantCohortAnnotation(null, null, null, null, "stest"), AnnotationType.ENUM);
-    }
-
-    @Test
-    public void deleteParticipantCohortAnnotation() throws Exception {
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(new Workspace());
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.findParticipantCohortStatus(cohortReviewId, participantId)).thenReturn(new ParticipantCohortStatus());
-        doNothing().when(cohortReviewService).deleteParticipantCohortAnnotation(1L, cohortReviewId, participantId);
-
-        reviewController.deleteParticipantCohortAnnotation(namespace, name, cohortId, cdrVersionId, participantId, 1L);
-
-        verify(cohortReviewService).findCohort(cohortId);
-        verify(cohortReviewService).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verify(cohortReviewService).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService).findParticipantCohortStatus(cohortReviewId, participantId);
-        verify(cohortReviewService).deleteParticipantCohortAnnotation(1L, cohortReviewId, participantId);
-        verifyNoMoreMockInteractions();
-    }
-
-    @Test
-    public void deleteParticipantCohortAnnotationNullAnnotationId() throws Exception {
-        try{
-            reviewController.deleteParticipantCohortAnnotation(namespace, name, cohortId, cdrVersionId, participantId, null);
-            fail("Should have thrown a BadRequestException!");
-        } catch (BadRequestException e) {
-            assertEquals("Invalid Request: Please provide a valid cohort annotation definition id.", e.getMessage());
-        }
-    }
-
-    @Test
-    public void updateParticipantCohortAnnotation() throws Exception {
-        long annotationId = 1;
-        long cohortReviewId = 1;
-        ModifyParticipantCohortAnnotationRequest request = new ModifyParticipantCohortAnnotationRequest();
-
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(new Workspace());
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.updateParticipantCohortAnnotation(annotationId, cohortReviewId, participantId, request)).thenReturn(new org.pmiops.workbench.db.model.ParticipantCohortAnnotation());
-
-        reviewController.updateParticipantCohortAnnotation(namespace, name, cohortId, cdrVersionId, participantId, 1L, request);
-
-        verify(cohortReviewService).findCohort(cohortId);
-        verify(cohortReviewService).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verify(cohortReviewService).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService).updateParticipantCohortAnnotation(annotationId, cohortReviewId, participantId, request);
-        verifyNoMoreMockInteractions();
-    }
-
-    @Test
-    public void getParticipantCohortAnnotations() throws Exception {
-        long cohortReviewId = 1;
-
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.READER)).thenReturn(new Workspace());
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.findParticipantCohortAnnotations(cohortReviewId, participantId)).thenReturn(new ArrayList<>());
-
-        reviewController.getParticipantCohortAnnotations(namespace, name, cohortId, cdrVersionId, participantId);
-
-        verify(cohortReviewService).findCohort(cohortId);
-        verify(cohortReviewService).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.READER);
-        verify(cohortReviewService).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService).findParticipantCohortAnnotations(cohortReviewId, participantId);
-        verifyNoMoreMockInteractions();
-    }
-
-    private ParticipantCohortAnnotation createParticipantCohortAnnotation(Boolean booleanValue, String stringValue, Integer integerValue, String dateValue, String enumValue) {
-        return new ParticipantCohortAnnotation()
-                .cohortAnnotationDefinitionId(1L)
-                .annotationId(1L)
-                .participantId(participantId)
-                .cohortReviewId(cohortReviewId)
-                .annotationValueBoolean(booleanValue)
-                .annotationValueString(stringValue)
-                .annotationValueInteger(integerValue)
-                .annotationValueDate(dateValue)
-                .annotationValueEnum(enumValue);
-    }
-
-    private org.pmiops.workbench.db.model.ParticipantCohortAnnotation createParticipantCohortAnnotation(ParticipantCohortAnnotation request) throws Exception {
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        Date date = request.getAnnotationValueDate() == null ? null : new Date(sdf.parse(request.getAnnotationValueDate().toString()).getTime());
-        CohortAnnotationEnumValue enumValue = null;
-        if (request.getAnnotationValueEnum() != null) {
-            enumValue = new CohortAnnotationEnumValue();
-            enumValue.setName(request.getAnnotationValueEnum());
-        }
-        return new org.pmiops.workbench.db.model.ParticipantCohortAnnotation()
-                .cohortAnnotationDefinitionId(request.getCohortAnnotationDefinitionId())
-                .annotationId(request.getAnnotationId())
-                .participantId(request.getParticipantId())
-                .cohortReviewId(request.getCohortReviewId())
-                .annotationValueBoolean(request.getAnnotationValueBoolean())
-                .annotationValueString(request.getAnnotationValueString())
-                .annotationValueInteger(request.getAnnotationValueInteger())
-                .annotationValueDate(date)
-                .annotationValueEnum(request.getAnnotationValueEnum())
-                .cohortAnnotationEnumValue(enumValue);
-    }
-
-    private void assertCreateParticipantCohortAnnotation(ParticipantCohortAnnotation request, AnnotationType annotationType) throws Exception {
-
-        ParticipantCohortStatusKey key = new ParticipantCohortStatusKey().cohortReviewId(cohortReviewId).participantId(participantId);
-        ParticipantCohortStatus participantCohortStatus = new ParticipantCohortStatus().participantKey(key);
-
-        CohortAnnotationDefinition cohortAnnotationDefinition = new CohortAnnotationDefinition().annotationType(annotationType);
-        if (request.getAnnotationValueEnum() != null) {
-            CohortAnnotationEnumValue cohortAnnotationEnumValue = new CohortAnnotationEnumValue().name(request.getAnnotationValueEnum());
-            cohortAnnotationDefinition.setEnumValues(new TreeSet(Arrays.asList(cohortAnnotationEnumValue)));
-        }
-        when(cohortReviewService.findCohort(cohortId)).thenReturn(createCohort(cohortId, workspaceId, null));
-        when(cohortReviewService.validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER)).thenReturn(new Workspace());
-        when(cohortReviewService.findCohortReview(cohortId, cdrVersionId)).thenReturn(createCohortReview(0, cohortId, cohortReviewId, cdrVersionId, null));
-        when(cohortReviewService.saveParticipantCohortAnnotation(isA(Long.class), isA(org.pmiops.workbench.db.model.ParticipantCohortAnnotation.class)))
-                .thenReturn(createParticipantCohortAnnotation(request));
-
-        ParticipantCohortAnnotation response =
-                reviewController.createParticipantCohortAnnotation(namespace, name, cohortId, cdrVersionId, participantId, request).getBody();
-
-        assertEquals(request, response);
-
-        verify(cohortReviewService, atLeastOnce()).findCohort(cohortId);
-        verify(cohortReviewService, atLeastOnce()).validateMatchingWorkspace(namespace, name, workspaceId, WorkspaceAccessLevel.WRITER);
-        verify(cohortReviewService, atLeastOnce()).findCohortReview(cohortId, cdrVersionId);
-        verify(cohortReviewService, atLeastOnce()).saveParticipantCohortAnnotation(isA(Long.class), isA(org.pmiops.workbench.db.model.ParticipantCohortAnnotation.class));
-        verifyNoMoreMockInteractions();
-    }
-
-    private Map<String, Map<Long, String>> createGenderRaceEthnicityConcept() {
-        Map<Long, String> raceMap = new HashMap<>();
-        raceMap.put(1L, "White");
-        raceMap.put(2L, "African American");
-
-        Map<Long, String> genderMap = new HashMap<>();
-        genderMap.put(3L, "MALE");
-        genderMap.put(4L, "FEMALE");
-
-        Map<Long, String> ethnicityMap = new HashMap<>();
-        ethnicityMap.put(5L, "Hispanic or Latino");
-        ethnicityMap.put(6L, "Not Hispanic or Latino");
-
-        Map<String, Map<Long, String>> concepts = new HashMap<>();
-        concepts.put(ParticipantCohortStatusColumns.RACE.name(), raceMap);
-        concepts.put(ParticipantCohortStatusColumns.GENDER.name(), genderMap);
-        concepts.put(ParticipantCohortStatusColumns.ETHNICITY.name(), ethnicityMap);
-        return concepts;
-    }
-
-    private ParticipantCohortStatus createParticipantCohortStatus(long cohortReviewId, long participantId, CohortStatus cohortStatus) {
-        ParticipantCohortStatusKey key = new ParticipantCohortStatusKey().cohortReviewId(cohortReviewId).participantId(participantId);
-        final Date dob = new Date(System.currentTimeMillis());
-
-        return new ParticipantCohortStatus()
-                .participantKey(key)
-                .status(cohortStatus)
-                .birthDate(dob)
-                .ethnicityConceptId(1L)
-                .genderConceptId(1L)
-                .raceConceptId(1L);
-    }
-
-    private CohortReview createCohortReview(long reviewSize, long cohortId, long cohortReviewId, long cdrVersionId, ReviewStatus reviewStatus) {
-        return new CohortReview()
-                .cohortId(cohortId)
-                .reviewSize(reviewSize)
-                .cohortReviewId(cohortReviewId)
-                .cdrVersionId(cdrVersionId)
-                .matchedParticipantCount(1000)
-                .creationTime(new Timestamp(System.currentTimeMillis()))
-                .reviewStatus(reviewStatus);
-    }
-
-    private Cohort createCohort(long cohortId, long workspaceId, String definition) {
-        Cohort cohort = new Cohort();
-        cohort.setCohortId(cohortId);
-        cohort.setWorkspaceId(workspaceId);
-        cohort.setCriteria(definition);
-        return cohort;
-    }
-
-    private Workspace createWorkspace(long workspaceId, String namespace, String name) {
-        Workspace workspace = new Workspace();
-        workspace.setWorkspaceId(workspaceId);
-        workspace.setWorkspaceNamespace(namespace);
-        workspace.setFirecloudName(name);
-        return workspace;
-    }
-
-    private void verifyNoMoreMockInteractions() {
-        verifyNoMoreInteractions(cohortReviewService, bigQueryService, participantCounter, domainLookupService, workspaceService);
-    }
-
+    return new org.pmiops.workbench.model.CohortReview()
+        .cohortReviewId(actualReview.getCohortReviewId())
+        .cohortId(actualReview.getCohortId())
+        .cdrVersionId(actualReview.getCdrVersionId())
+        .creationTime(actualReview.getCreationTime().toString())
+        .matchedParticipantCount(actualReview.getMatchedParticipantCount())
+        .reviewSize(actualReview.getReviewSize())
+        .reviewedCount(actualReview.getReviewedCount())
+        .participantCohortStatuses(newParticipantCohortStatusList)
+        .page(page)
+        .pageSize(pageSize)
+        .sortOrder(sortOrder.toString())
+        .sortColumn(sortColumn.name());
+  }
 }
